accept JobConfig and PackageConfig equally
Since we made JobConfig to be as powerful as PackageConfig, we need to
make sure that packit works equally well with JobConfig as with
PackageConfig.

Hunor suggested to type-annotate api.package_config as
CommonPackageConfig so we would make sure that only the common
attributes of JobConfig and PackageConfig would be utilized.

Follow-up to: #858
Needed for packit-service/packit-service#418
Needed for packit-service/packit-service#676
